Make Vectronic compile with new steam_api

+++ src/game/client/vectronic/GameUI/vhybridbutton.cpp
@@ -947,7 +947,7 @@ void BaseModHybridButton::OnKeyCodePressed( vgui::KeyCode code )
 	BaseModUI::CBaseModPanel::GetSingleton().SetLastActiveUserId( iJoystick );
 
 	int iController = XBX_GetUserId( iJoystick );
-	bool bIsPrimaryUser = ( iController >= 0 && XBX_GetActiveUserId() == DWORD( iController ) );
+	bool bIsPrimaryUser = ( iController >= 0 && DWORD(XBX_GetActiveUserId()) == DWORD( iController ) );
 
 	KeyCode localCode = GetBaseButtonCode( code );
